refactor(suite-common): split stablecoin-yield folder into subfolders
