+++ suite-common/wallet-core/src/index.ts
@@ -70,16 +70,16 @@ export * from './stake/stakeSelectors';
 export * from './stake/stakeThunks';
 export * from './stablecoin-yield/stablecoinYieldReducer';
 export * from './stablecoin-yield/stablecoinYieldSelectors';
-export * from './stablecoin-yield/stablecoinYieldApprovalActionUtils';
-export * from './stablecoin-yield/stablecoinYieldApprovalThunks';
+export * from './stablecoin-yield/utils/stablecoinYieldApprovalActionUtils';
+export * from './stablecoin-yield/thunks/stablecoinYieldApprovalThunks';
 export * from './stablecoin-yield/stablecoinYieldConstants';
-export * from './stablecoin-yield/stablecoinYieldDepositThunks';
-export * from './stablecoin-yield/stablecoinYieldDeviceUtils';
-export * from './stablecoin-yield/stablecoinYieldFeeEstimation';
+export * from './stablecoin-yield/thunks/stablecoinYieldDepositThunks';
+export * from './stablecoin-yield/utils/stablecoinYieldDeviceUtils';
+export * from './stablecoin-yield/utils/stablecoinYieldFeeEstimation';
 export * from './stablecoin-yield/stablecoinYieldTypes';
-export * from './stablecoin-yield/stablecoinYieldUtils';
-export * from './stablecoin-yield/stablecoinYieldWrapThunks';
-export * from './stablecoin-yield/useWrappedNativePendingTx';
+export * from './stablecoin-yield/utils/stablecoinYieldUtils';
+export * from './stablecoin-yield/thunks/stablecoinYieldWrapThunks';
+export * from './stablecoin-yield/hooks/useWrappedNativePendingTx';
 export * from './stake/tron/tronStakeReducer';
 export * from './stake/tron/tronStakeSelectors';
 export { composeTronFreezeFeeLevelsThunk } from './stake/tron/actions/freeze/composeFreeze';
