+++ suite-common/wallet-core/src/stablecoin-yield/stablecoinYieldSelectors.ts
@@ -1,6 +1,5 @@
 import { type DeviceRootState, selectSelectedDevice } from '@suite-common/device';
 
-import { isWrappedNativeFlowSupported } from './stablecoinYieldDeviceUtils';
 import {
     type StablecoinYieldRootState,
     type StablecoinYieldSessionState,
@@ -9,6 +8,7 @@ import {
     initialStablecoinYieldSessionState,
 } from './stablecoinYieldReducer';
 import type { YieldFlowType } from './stablecoinYieldTypes';
+import { isWrappedNativeFlowSupported } from './utils/stablecoinYieldDeviceUtils';
 
 export const selectIsWrappedNativeFlowSupported = (state: DeviceRootState): boolean =>
     isWrappedNativeFlowSupported(selectSelectedDevice(state));
