+++ suite-common/wallet-core/src/stablecoin-yield/utils/stablecoinYieldApprovalActionUtils.ts
