+++ suite-common/wallet-core/src/stablecoin-yield/utils/stablecoinYieldDeviceUtils.test.ts
