+++ suite-common/wallet-core/src/stablecoin-yield/utils/stablecoinYieldFeeEstimation.test.ts
